+++ kotonebot/core/__init__.py
@@ -1,13 +1,13 @@
 from .entities.base import Prefab
-from .entities.template_match import TemplateMatchPrefab, TemplateMatchFindKargs
-from .entities.ocr import OcrPrefab, OcrFindKargs
+from .entities.template_match import TemplateMatchPrefab, TemplateMatchQuery
+from .entities.ocr import OcrPrefab
 from .entities.base import GameObject
 from .entities.base import GameObjectType
 from .entities.compound import AnyOf
 from .bot import *  # noqa: F403
 
 __all__ = [
-    'Prefab', 'TemplateMatchPrefab', 'OcrPrefab',
+    'Prefab', 'TemplateMatchPrefab', 'OcrPrefab', 'TemplateMatchQuery',
     'GameObject',
     'GameObjectType',
     'AnyOf'